+++ NEWS.md
@@ -1,9 +1,12 @@
-## 2019-05-17 v1.3.0-SNAPSHOT
+## 2019-06-12 v1.3.0
 * Updated association profiles samples. 
-* Updated API endpoint for getting details by master id. Added query param to specify return type. Added searching by cql query for details
-* Updated API endpoint for getting masters by detail id. Added query param to specify return type. Added searching by cql query for details
+* Updated API endpoint for getting details by master id. Added query param to specify return type. Added searching by cql query for details.
+* Updated API endpoint for getting masters by detail id. Added query param to specify return type. Added searching by cql query for details.
 * Updated API endpoints for Profile Association entity. Added query parameters to specify master and detail types of profile association.
 * Updated Action Profile sample data.
+* Added check for existence of associated profiles before delete.
+* Added validation on post and put for action, mapping and match profiles.
+* Fixed 'name' field validation.
 
   | METHOD |             URL                                              | DESCRIPTION                         |
   |--------|--------------------------------------------------------------|-------------------------------------|